+++ tests/serving/test_tracking.py
@@ -14,10 +14,14 @@
 #
 import json
 from pprint import pprint
+from unittest.mock import patch
 
 import numpy as np
+import pytest
 
 import mlrun
+from mlrun.common.schemas import ModelMonitoringMode
+from tests.serving.test_serving import _log_model
 
 testdata = '{"inputs": [[5, 6]]}'
 
@@ -91,6 +95,35 @@ def test_ensemble_tracking():
         "2": ["ModelTestingClass", [[5, 6]], [15]],
         "VotingEnsemble": ["VotingEnsemble", [[5, 6]], [12.5]],
     }
+
+
+@pytest.mark.parametrize("enable_tracking", [False])
+def test_tracked_function(rundb_mock, enable_tracking):
+    with patch("mlrun.get_run_db", return_value=rundb_mock):
+        project = mlrun.new_project("test-pro", save=False)
+        fn = mlrun.new_function("test-fn", kind="serving")
+        model_uri = _log_model(project)
+        print(model_uri)
+        fn.add_model("m1", model_uri, "ModelTestingClass", multiplier=5)
+        fn.set_tracking("dummy://", enable_tracking=enable_tracking)
+        server = fn.to_mock_server(track_models=True)
+        server.test("/v2/models/m1/infer", testdata)
+
+        dummy_stream = server.context.stream.output_stream
+        assert len(dummy_stream.event_list) == 1, "expected stream to get one message"
+
+    rundb_mock.patch_model_endpoint.assert_called_once()
+    assert (
+        rundb_mock.patch_model_endpoint.call_args.kwargs["attributes"]["model_uri"]
+        == model_uri
+    ), "model_uri attribute of the model endpoint was not updated as expected"
+    if not enable_tracking:
+        assert (
+            rundb_mock.patch_model_endpoint.call_args.kwargs["attributes"][
+                "monitoring_mode"
+            ]
+            == ModelMonitoringMode.disabled
+        ), "model_uri attribute of the model endpoint was not updated as expected"
 
 
 def rec_to_data(rec):
